Assignment 8/requirements.txt
pandas
numpy
streamlit
langchain
openai
chromadb
sentence-transformers
faiss-cpu
python-dotenv
